requirements.txt
discord.py==2.*
requests==2.*
